Organize components directory with modular structure
- **Modularize Components**: Move all standalone components to directories
  - FloatingShowButton: Create directory structure with index.ts
  - HeaderBar: Create directory structure with index.ts
  - OverlayUI: Create directory structure with index.ts
  - PWAUpdateNotification: Create directory structure with index.ts

- **Fix Import Paths**: Update relative imports for new directory structure
  - Fix ui/Button imports (../ui/Button)
  - Fix types/settings imports (../../types/settings)
  - Update OverlayUI component imports to use parent directories

- **Maintain Consistency**: All components now follow unified structure
  - Component file + Test file + index.ts export
  - Consistent with existing FooterBar, PdfViewer, SettingsPanel modules

- **Quality Assurance**:
  - ✅ 94/94 tests pass across 6 test files
  - ✅ TypeScript compilation successful
  - ✅ All component functionality preserved

Improved maintainability and consistency across component architecture.
