+++ src/components/FloatingShowButton/index.ts
@@ -1,0 +1,1 @@
+export { FloatingShowButton } from './FloatingShowButton';
