+++ src/components/PWAUpdateNotification/index.ts
@@ -1,0 +1,1 @@
+export { PWAUpdateNotification } from './PWAUpdateNotification';
